+++ src/estimator/add_policy_estimator.rs
@@ -211,7 +211,7 @@ pub(super) fn estimate_add_policy(token_blocks: &[DeflateTokenBlock]) -> Diction
 fn verify_miniz1_recognition() {
     use crate::deflate::deflate_reader::parse_deflate;
 
-    let v = crate::process::read_file("compressed_minizoxide_level1.deflate");
+    let v = crate::utils::read_file("compressed_minizoxide_level1.deflate");
 
     let contents = parse_deflate(&v).unwrap();
 
@@ -232,7 +232,7 @@ fn verify_zlib_level_recognition() {
     ];
 
     for i in 1..=4 {
-        let v = crate::process::read_file(&format!("compressed_zlib_level{}.deflate", i));
+        let v = crate::utils::read_file(&format!("compressed_zlib_level{}.deflate", i));
 
         let contents = parse_deflate(&v).unwrap();
         let add_policy = estimate_add_policy(&contents.blocks);
@@ -252,7 +252,7 @@ fn verify_zlibng_level_recognition() {
     ];
 
     for i in 1..=4 {
-        let v = crate::process::read_file(&format!("compressed_zlibng_level{}.deflate", i));
+        let v = crate::utils::read_file(&format!("compressed_zlibng_level{}.deflate", i));
 
         let contents = parse_deflate(&v).unwrap();
         let add_policy = estimate_add_policy(&contents.blocks);
@@ -267,7 +267,7 @@ fn verify_libdeflate_level_recognition() {
     use crate::deflate::deflate_reader::parse_deflate;
 
     for i in 1..=9 {
-        let v = crate::process::read_file(&format!("compressed_libdeflate_level{}.deflate", i));
+        let v = crate::utils::read_file(&format!("compressed_libdeflate_level{}.deflate", i));
 
         let contents = parse_deflate(&v).unwrap();
         let add_policy = estimate_add_policy(&contents.blocks);
