+++ src/utils.rs
@@ -1,0 +1,24 @@
+#[allow(dead_code)]
+#[cfg(test)]
+pub fn write_file(filename: &str, data: &[u8]) {
+    let mut writecomp = std::fs::File::create(filename).unwrap();
+    std::io::Write::write_all(&mut writecomp, data).unwrap();
+}
+
+#[cfg(test)]
+pub fn read_file(filename: &str) -> Vec<u8> {
+    use std::fs::File;
+    use std::io::Read;
+    use std::path::Path;
+
+    let filename = Path::new(env!("CARGO_MANIFEST_DIR"))
+        .join("samples")
+        .join(filename);
+    println!("reading {0}", filename.to_str().unwrap());
+    let mut f = File::open(filename).unwrap();
+
+    let mut content = Vec::new();
+    f.read_to_end(&mut content).unwrap();
+
+    content
+}
